Do more explicit casting from size_t to uint
GIT_SILENT

+++ kdevplatform/util/tests/test_embeddedfreetree.cpp
@@ -293,7 +293,7 @@ public:
 
     uint size() const
     {
-        return realSet.size();
+        return static_cast<uint>(realSet.size());
     }
 
     uint getItem(uint number) const
